standard: Install software from env file instead of tarball
Using an environment file has the benefit of having a list of the
installed packages explicitly documented inside the repo, instead of
just adding a magic tarball from an outside system. The environment is
usually resolved using our internal builder tool (based on the
[production] and [dev] config), but in theory the file could be
generated with any conda/mamba installation. It currently takes 1-2 hrs
to build the environment and less than 10 minutes to install it into the
image.

Also removed the install commands for packages [jsonpickle] and
[vaderSentiment] which have been added to the build environment
directly.

[production]: https://github.com/AaltoSciComp/science-build-configs/blob/master/configs/aalto-jupyter-anaconda/anaconda/build_config.yaml
[dev]: https://github.com/AaltoSciComp/science-build-configs/blob/master/configs/aalto-jupyter-anaconda/anaconda/build_config.yaml
[jsonpickle]: https://github.com/AaltoSciComp/science-build-configs/commit/b3e89fc43f1f1bdcf6b13203339bf0ce2546354d
[vaderSentiment]: https://github.com/AaltoSciComp/science-build-configs/commit/5255b4a68b2080fc93dd0980b4f93434a5455209