Remove conda package cache
The package cache directory (pgks_dir, ~/.conda/pkgs) is used to unpack
downloaded packages, which are then hardlinked to the environment
directory. Deleting the cache is safe, because the hard link inside the
environment still stays. The conda setting `allow_softlinks` is set to
ensure that conda either copies the files or creates hardlinks instead
of symlinking installed packages from the cache directory.