Use a registry to cache builds
Creates a buildx docker-container driver [1] to send the build cache to and
receive from GitHub. Harbor cannot currently be used to store the cache
because of a bug [2].

Because the docker-container driver is isolated from the system docker
and can't access the default image store, the base image has to be
pushed into a registry before building other images.

[1] https://docs.docker.com/build/drivers/docker-container/
[2] https://github.com/goharbor/harbor/issues/14774